Improved: null check in consume FulfillmentFeed service to verify against the file content instead of messageText
This is done because of the change done in poll#SystemMessageFileSftp service, the messageText will now hold the file path instead of the file content, so removed null check on fileText, instead we need to check the expected shipments data being read from the filePath.
If no shipments, one scenario is transformation in NiFi generating file with null value as no valid shipments data, or it may be due to some other unexpected JSON generated to be consumed, then error will be logged and can be further checked for a possible issue.